cleaned up code for feed by adding separate components for post, comment, reply, and addpost
